% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations procedures.R
\name{iteration}
\alias{iteration}
\title{Iteration procedure of Near Neutral Communities}
\usage{
iteration(comm, breaks = 10000)
}
\arguments{
\item{comm}{Object of class pool or local}

\item{breaks}{Maximum number of iteration}
}
\value{
Iteration function return the original class object after run several iterations of death and birth events, such events should be accessed by attributes,see example.
}
\description{
Function to run a iteration procedure in Local and Pool Communities
}
\examples{
setseed(2024)
mypool<-pool(J=1000,theta = 100,sigma=0.03)
mypool #see the species pool
attr(mypool,"richness") #only one richness recorded

mynewpool<-iteration(mypool)
mynewpool
attr(mynewpool,"richness") #a set of richness recorded
}
\seealso{
pool one.local metacomm
}
