% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanPool.R
\name{meanPool}
\alias{meanPool}
\title{Calculate the relative species abundance and W of the pool.}
\usage{
meanPool(df)
}
\arguments{
\item{df}{A dataframe containing the results of the simulations.}
}
\value{
The meanPool function returns an object containing the species rank, RSA (Relative Species Abundance), RSA standard deviation, W, W standard deviation and the simulation's sigma value.
}
\description{
Function to extract the arithmetic mean of the RSA and W from all simulations performed by the modelMetacomm function.
}
\examples{
mymodel <- modelMetacomm(J=100,theta = 10,sigma=0.03,Jl=10,nloc=3,mig=0.25,loccomm="diversity",abonly=F,nSim=2)
results <- meanPool(mymodel)
}
