% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetacommunity.R
\name{metacomm}
\alias{metacomm}
\title{Create a Metacommunity in Near Neutral Communities Models}
\usage{
metacomm(
  J,
  theta,
  sigma,
  Jl,
  nloc,
  mig,
  loccomm = "pool",
  abonly = F,
  dataFrame = T
)
}
\arguments{
\item{J}{The number of individuals in Species Pool}

\item{theta}{Hubbels Fundamental Number of Biodiversity}

\item{sigma}{the standard deviation for change the per capita fecundity parameter w}

\item{Jl}{The number of individuals in local species}

\item{nloc}{Number of local communities}

\item{mig}{Migration rate number}

\item{loccomm}{Option of get the per capita fecundity parameter in local communities: "pool", "uniform", "diversity" or "neutral"}

\item{abonly}{Logical Use only abundance of species pool in migration events instead include per capita fecundity parameter w}

\item{dataframe}{Logical output as data.frame and metacommunity object instead list object}
}
\value{
metacomm function return a metacomm object or a data.frame with information of a set of local communities.

Such Local communities could attributes a set of wi (e.g. loccom="diversity") or get a uniform wi for all local communities (e.g. loccom="diversity"), or keep the wi of species pool (e.g. loccom="pool").

Also Local communities could be neutral if loccom="neutral" or if sigma = 0, in this options all metacommunity will be neutral.
}
\description{
Function to run the metacommunity to model in Near Neutral Metacommunities
}
\examples{
#Keep the wi of specie pool
set.seed(2024)
myMeta<-metacomm(J=1000,theta = 100,sigma=0.03,Jl=100,nloc=3,mig=0.25,loccomm="pool",abonly=F,dataFrame=T)
myMeta #see the species pool
tapply(myMeta$wi,list(myMeta$species,myMeta$source),sum)

# Uniform wi in Local community
set.seed(2024)
myMeta<-metacomm(J=1000,theta = 100,sigma=0.03,Jl=100,nloc=3,mig=0.25,loccomm="uniform",abonly=F,dataFrame=T)
myMeta #see the species pool
tapply(myMeta$wi,list(myMeta$species,myMeta$source),sum)

# Diversity wi in Local community
set.seed(2024)
myMeta<-metacomm(J=1000,theta = 100,sigma=0.03,Jl=100,nloc=3,mig=0.25,loccomm="diversity",abonly=F,dataFrame=T)
myMeta #see the species pool
tapply(myMeta$wi,list(myMeta$species,myMeta$source),sum)

# Neutral Local community
set.seed(2024)
myMeta<-metacomm(J=1000,theta = 100,sigma=0.03,Jl=100,nloc=3,mig=0.25,loccomm="neutral",abonly=F,dataFrame=T)
myMeta #see the species pool
tapply(myMeta$wi,list(myMeta$species,myMeta$source),sum)

# Neutral metacommunity
set.seed(2024)
myMeta<-metacomm(J=1000,theta = 100,sigma=0,Jl=100,nloc=3,mig=0.25,loccomm="diversity",abonly=F,dataFrame=T)
myMeta #see the species pool
tapply(myMeta$wi,list(myMeta$species,myMeta$source),sum)
}
\seealso{
iteration pool metacomm
}
