% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMetacom.R
\name{modelMetacomm}
\alias{modelMetacomm}
\title{Model a Metacommunity in Near Neutral Communities}
\usage{
modelMetacomm(
  J,
  theta,
  sigma,
  Jl,
  nloc,
  mig,
  loccomm = "pool",
  abonly = F,
  nSim = 5,
  file = NULL,
  exp = T,
  itera = T
)
}
\arguments{
\item{J}{The number of individuals in Species Pool}

\item{theta}{Hubbels Fundamental Number of Biodiversity}

\item{sigma}{the standard deviation for change the per capita fecundity parameter w}

\item{Jl}{The number of individuals in local species}

\item{nloc}{Number of local communities}

\item{mig}{Migration rate number}

\item{loccomm}{Option of get the per capita fecundity parameter in local communities: "pool", "uniform", "diversity" or "neutral"}

\item{abonly}{Logical Use only abundance of species pool in migration events instead include per capita fecundity parameter w}

\item{nSim}{Number of simulations of the model}

\item{file}{Optional file to exporte the output to *.rData file}
}
\value{
The model of metacommunity
}
\description{
Function to run the model of Near Neutral Metacommunities
}
\examples{
#Get an object in R
set.seed(2024)

myModel<-modelMetacomm(J=1000,theta = 100,sigma=0.03,Jl=100,nloc=3,mig=0.25,loccomm="diversity",abonly=F,nSim=2)
myModel #see the species pool

#Save in external file
set.seed(2024)
modelMetacomm(J=1000,theta = 100,sigma=0.03,Jl=100,nloc=3,mig=0.25,loccomm="diversity",abonly=F,nSim=2,file="mytest.RData")
load("mytest.RData") #see the species pool
model_1000_100_0.03_100_3_0.25_diversity
}
\seealso{
iteration pool metacomm one.local
}
