% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLocal.R
\name{one.local}
\alias{one.local}
\title{Create a local community in Near Neutral Communities Models}
\usage{
one.local(pool, jl, mig, sigma = NULL, wi = NULL, abonly = F)
}
\arguments{
\item{pool}{An object of class pool}

\item{jl}{The number of individuals in local species}

\item{mig}{Migration rate number}

\item{sigma}{the standard deviation for change the per capita fecundity parameter w}

\item{wi}{Vector of wi to impose uniformity in wi for a set of local communities}

\item{abonly}{Logical Use only abundance of species pool in migration events instead include per capita fecundity parameter w}
}
\value{
pool function return a local class object with Species label, Abundance and per capita fecundity parameter w. Such object also record other parameters in attributes
}
\description{
Function to run the local community of species to model in Near Neutral Metacommunities
}
\examples{
set.seed(2024)
mypool<-pool(J=1000,theta = 100,sigma=0.03)
mypool #see the species pool
attributes(mypool)
attr(mypool,"richness") #only one richness recorded

mylocal<-one.local(mypool,jl=100,mig=0.3,sigma=0.02,wi=NULL,abonly=F)
attributes(mylocal)
attr(mylocal,"richness")
}
\seealso{
iteration pool metacomm
}
