% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePool.R
\name{pool}
\alias{pool}
\title{Create a Pool of Species in Near Neutral Communities Models}
\usage{
pool(J, theta, sigma, ord = T)
}
\arguments{
\item{J}{The number of individuals in Species Pool}

\item{theta}{Hubbels Fundamental Number of Biodiversity}

\item{sigma}{the standard deviation for change the per capita fecundity parameter w}

\item{ord}{Logical to make the decreasing ordination of abundance in species pool}
}
\value{
pool function return a pool class object with Species label, Abundance and per capita fecundity parameter w. Such object also record richness and initial parameters in attributes
}
\description{
Function to run the pool of species to model in Near Neutral Metacommunities
}
\examples{
set.seed(2024)
mypool<-pool(J=1000,theta = 100,sigma=0.03)
mypool #see the species pool
attributes(mypool)
attr(mypool,"richness") #only one richness recorded

mynewpool<-iteration(mypool)
mynewpool
attr(mynewpool,"richness") #a set of richness recorded
}
\seealso{
iteration one.local metacomm
}
